// src/app/interfaces/forms/reporte-de-fugas.ts
export interface ReporteDeFugas {
  formId: number;
  userSubscriptionCode?: string;
  identificationType: string;
  identification: number;
  name1: string;
  name2: string;
  lastName1: string;
  lastName2: string;
  neighborhood: string;
  address: string;
  mobile: string;
  phoneNumber: string;
  typeOfLeakId: number;
  email: string;
  politicId: number;
  recaptchaToken: string;
}